backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
pydantic==2.8.2
pandas==2.2.2
numpy==1.26.4
sqlalchemy==2.0.32
alembic==1.13.2
scikit-learn==1.5.1
python-dotenv==1.0.1
pytesseract==0.3.10
Pillow==10.3.0
pdf2image==1.17.0
